📝 Zero Values

+++ src/04_Zero-Values/main.go
@@ -3,8 +3,9 @@ package main
 import "fmt"
 
 func main() {
+
 	// Zero values
-	var a uint8   // Default = 0
+	var a int     // Default = 0
 	var b float64 // Default = 0
 	var c string  // String Vacio
 	var d bool    // Default = False
